+++ include/vrv/vrvdef.h
@@ -174,6 +174,7 @@ enum ClassId {
     DOTS,
     FLAG,
     FTREM,
+    GRACEGRP,
     HALFMRPT,
     KEYSIG,
     KEYACCID,